refactor: move config.ts into src
